Reduce index size by capping term count

+++ posting.h
@@ -6,13 +6,13 @@
 struct posting {
 	size_t id;
 	size_t diff;
-	uint16_t count;
+	uint8_t count;
 	size_t id_capacity;
 	size_t id_length;
 	uint8_t *id_store;
 	size_t count_capacity;
 	size_t count_length;
-	uint16_t *count_store;
+	uint8_t *count_store;
 };
 
 struct posting *posting_new();
